requirements.txt
flask==2.3.3
python-dotenv==1.0.0
firebase-admin==6.2.0
flask-wtf==1.1.1
gunicorn==21.2.0